transform_solutions for time-dependent results
